+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to insta and cargo-insta are documented here.
 ## 1.12.0
 
 - Add support for sorting redactions (`sorted_redaction` and `Settings::sort_selector`). (#212)
+- Changed snapshot name detection to no longer use thread names but function names. (#213)
 
 ## 1.11.0